Add a progress spinner to show that the Episode image is being loaded.
While the image is being loaded, the space reserved for the image at the view is kept gray. After the image is loaded, the spinner disappears and the image is shown. If there is not an image for the episode or it is not available, a generic image is shown.