Starts implementing the LruRepositoryManager.
It will be used as policy manager for the episode images directory (repository).

It has not yet being used where it should be. It should be instantiated in the ImageDirectory and used as a decorator
for episodeDirectory. Then, EpisodeImagesCache should be removed from that class.

The calculation of the parameter that should be used in to instatiate the LruRepositoryManager is in its implementation
maked with a "to be removed" TODO.