Use ConcurrentSkipListMap instead of TreeMap in Season and SeasonSet.
To solve concurrency problems in Season and SeasonSet when adding
episodes from multiple threads, I replaced the TreeMaps in ...domain.model.Season
and ...domain.model.SeasonSet, as suggested in the Concurrent Collections section
of https://developer.android.com/reference/java/util/concurrent/package-summary.html.

The number of download threads was also fixed to '6' (because it's a cool number)
instead of the number of seasons, for consistency.